requirements.txt
Django==1.8.2
django-admin-bootstrapped==2.5.0
django-autocomplete-light==2.2.0rc4
psycopg2==2.6
pytz==2015.4
djangorestframework==3.1.2
nltk==3.0
django-widget-tweaks
django-recaptcha
mock
django-allauth
django-fluent-comments
django-contrib-comments
-e git+https://github.com/maraujop/django-crispy-forms.git@dev#egg=django-crispy-forms
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.3.0
static3==0.6.1
wsgiref==0.1.2
